#713 move Link and LinkConstants to separate commons.link package make Link.getHtmlAttributes non-null returning an empty map for invalid links cosmetic changes in javadocs and file headers
