+++ bundles/core/src/main/java/com/adobe/cq/wcm/core/components/commons/link/package-info.java
@@ -13,36 +13,10 @@
  ~ See the License for the specific language governing permissions and
  ~ limitations under the License.
  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
-package com.adobe.cq.wcm.core.components.internal;
+/**
+ * Link handling.
+ */
+@Version("1.0.0")
+package com.adobe.cq.wcm.core.components.commons.link;
 
-import org.apache.commons.lang3.StringUtils;
-
-public enum Heading {
-
-    H1("h1"),
-    H2("h2"),
-    H3("h3"),
-    H4("h4"),
-    H5("h5"),
-    H6("h6");
-
-    private String element;
-
-    Heading(String element) {
-        this.element = element;
-    }
-
-    public static Heading getHeading(String value) {
-        for (Heading heading : values()) {
-            if (StringUtils.equalsIgnoreCase(heading.element, value)) {
-                return heading;
-            }
-        }
-        return null;
-    }
-
-    public String getElement() {
-        return element;
-    }
-
-}
+import org.osgi.annotation.versioning.Version;
