+++ bundles/core/src/test/java/com/adobe/cq/wcm/core/components/internal/link/LinkImplTest.java
@@ -17,13 +17,13 @@ package com.adobe.cq.wcm.core.components.internal.link;
 
 import static com.adobe.cq.wcm.core.components.internal.link.LinkTestUtils.assertInvalidLink;
 import static com.adobe.cq.wcm.core.components.internal.link.LinkTestUtils.assertValidLink;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertSame;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.mockito.Mockito.mock;
 
 import org.junit.jupiter.api.Test;
 
-import com.adobe.cq.wcm.core.components.models.Link;
+import com.adobe.cq.wcm.core.components.commons.link.Link;
 import com.day.cq.wcm.api.Page;
 
 class LinkImplTest {
